% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_label.R
\name{meta_label}
\alias{meta_label}
\title{Attach labels from YAML metadata to a data frame.}
\usage{
meta_label(data, yaml_path, warn_missing = FALSE)
}
\arguments{
\item{data}{A data frame you want to label.}

\item{yaml_path}{Path to YAML file containing variable labels.}

\item{warn_missing}{Logical, controls warnings for metadata variables not found in data frame.}
}
\value{
Original data frame with labels attached
}
\description{
Reads variable labels from a YAML file and applies them to the corresponding columns of a data frame.
Data frame may contain variables not found in metadata, and vice versa.
}
\examples{
\dontrun{
yaml_path <- system.file("extdata", "codebook.yaml", package = "wcmtools")

df <- data.frame(
  age = c(23, 25, 42, 38, 29),
  gender = c(1, 2, 1, 1, 2),
  education = c(2, 3, 3, 5, 4)
)

labeled_df <- meta_label(df, yaml_path)
labelled::var_label(labeled_df$age)
}
}
